reestruturação do projeto, refactore na obtenção dos dados e no treinamento do modelo.
