+++ backend/metric_reports/model_metric_report_2025_07_05_12_31_56.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="pt-br">
+<head><meta charset="UTF-8"><title>Relatório de Métricas - Modelo</title>
+<style>body { font-family: Arial; background-color: #f9f9f9; margin: 40px; color: #333; }
+.metric-box { background: #e6f2ff; padding: 10px; border-left: 5px solid #005a9c; margin-bottom: 10px; }
+table { border-collapse: collapse; width: 100%; } th, td { border: 1px solid #ccc; padding: 8px; text-align: center; }</style>
+</head><body>
+<h1>Relatório de Métricas do Modelo</h1><p>Data: 2025_07_05_12_31_56</p>
+<div class="metric-box"><strong>Acurácia:</strong> 0.9256</div>
+<div class="metric-box"><strong>ROC AUC:</strong> 0.5000</div>
+<h2>Relatório de Classificação</h2>
+<table><tr><th>Classe</th><th>Precisão</th><th>Revocação</th><th>F1-score</th><th>Suporte</th></tr><tr><td>0</td><td>0.93</td><td>1.00</td><td>0.96</td><td>26890.0</td></tr><tr><td>1</td><td>0.00</td><td>0.00</td><td>0.00</td><td>2161.0</td></tr></table><h2>Matriz de Confusão</h2>
+<img src="confusion_matrix.png" alt="Matriz de Confusão" width="400"/>
+<hr/><p>Relatório gerado automaticamente pelo Decision AI.</p></body></html>
